+++ src/types/shims.d.ts
@@ -1,4 +1,0 @@
-declare module "markdown-it";
-declare module "markdown-it/lib/token.mjs";
-declare module "gray-matter";
-declare module "markdown-it-mark";